Adapt `gardenlet` to take over management of `machine-controller-manager` deployment (#8018)
* Add new feature gate to `gardenlet`


* Prepare extensions library for MCM management changes

- added a `--gardenlet-manages-mcm` flag to `GeneralOptions` which can be used in extensions
- added a new `mcmManaged` bool value to generic `Worker` actuator (not yet used; will follow in subsequent commits)
- added new `EnsureMachineControllerManager{Deployment,VPA}` methods to the generic `ControlPlane` ensurer (must be implemented by extensions; will follow in subsequent commits)


* Adapt generic `Worker` actuator

- For existing clusters, we must ensure that the old `ManagedResource` containing the RBAC privileges for MCM gets deleted in case `gardenlet` manages MCM. `gardenlet` deploys its own `ManagedResource` for these permissions.
- After the `Worker` restoration but before the `Worker` reconciliation, we have to make sure MCM gets scaled up. `gardenlet` is not aware of when the restoration is finished, so it has no control over it.


* Provide utility functions for MCM webhook injections


* `gardenlet`'s `Seed` controller deploys MCM resources

Similar to the CA, these contain a `ClusterRole` with the needed permission for MCM


* Initialize MCM component on `Shoot` object in botanist

Replica count determination logic copied from https://github.com/gardener/gardener/blob/master/extensions/pkg/controller/worker/genericactuator/actuator_reconcile.go#L64-L82


* `gardenlet`'s `Shoot` controller manages MCM


* `gardenlet`'s `Shoot` controller deploys MCM dashboard for Plutono

Dashboard copied from https://github.com/gardener/gardener/blob/master/charts/gardener/provider-local/internal/machine-controller-manager/seed/mcm-monitoring-dashboard.json


* Adapt documentation

* Address PR review feedback

---------

+++ extensions/pkg/controller/worker/genericactuator/actuator_migrate.go
@@ -42,13 +42,15 @@ func (a *genericActuator) Migrate(ctx context.Context, log logr.Logger, worker *
 		return fmt.Errorf("could not keep objects of managed resource containing mcm chart for worker '%s': %w", kubernetesutils.ObjectName(worker), err)
 	}
 
-	// Make sure machine-controller-manager is deleted before deleting the machines.
-	if err := a.deleteMachineControllerManager(ctx, log, worker); err != nil {
-		return fmt.Errorf("failed deleting machine-controller-manager: %w", err)
-	}
+	if a.mcmManaged {
+		// Make sure machine-controller-manager is deleted before deleting the machines.
+		if err := a.deleteMachineControllerManager(ctx, log, worker); err != nil {
+			return fmt.Errorf("failed deleting machine-controller-manager: %w", err)
+		}
 
-	if err := a.waitUntilMachineControllerManagerIsDeleted(ctx, log, worker.Namespace); err != nil {
-		return fmt.Errorf("failed deleting machine-controller-manager: %w", err)
+		if err := a.waitUntilMachineControllerManagerIsDeleted(ctx, log, worker.Namespace); err != nil {
+			return fmt.Errorf("failed deleting machine-controller-manager: %w", err)
+		}
 	}
 
 	if err := a.shallowDeleteAllObjects(ctx, log, worker.Namespace, &machinev1alpha1.MachineList{}); err != nil {
